events: use 'this' instead of global 'events'
This change to the events helper generator creates event wrappers
that call `this.on` instead of `events.on`. This should make no
difference to callers of `events.blockDestroy(...)` since `this`
is bound to `events` in that case.

However, it opens the door to intelligent "subclassing" of the
`events` object. For example, here is a function that creates a
"filtered" version of the `events` object according to some
predicate function, that will only report events that pass some
test:

```
function filterEvents(events, eventFilter) {
    var filteredEvents = Object.create(events);

    filteredEvents.on = function (eventType, handler, priority) {
        var filteredHandler = function (event) {
            if (eventFilter(event)) {
                return handler(event);
            }
        };
        events.on.call(this, eventType, filteredHandler, priority);
    };

    return filteredEvents;
}
```

This could be used to e.g. restrict an event handler to an area or a
given player.

```
var bobsEvents = filterEvents(events, function (e) {
  return e.player && e.player.name == 'Bob';
});

bobsEvents.playerArmSwing(function (e) {
  // will only be called when Bob swings his arm
});
```