Basic CoffeeScript support
Only in /js for now, and just tries to interpret everything as
CoffeeScript, falling back to JS if this fails (which fails to account
for an user trying to write CoffeeScript but making an error).

Also, works only on Nashorn due to usage of `double` as a property name
in the CoffeeScript compiler.