[1.13] Set api-version in plugin.yml to bypass compatibility layer
Spigot's automatic compatibility layer, which rewrites obsolete
calls to use CraftEvil and friends, is triggered by the absence
of an api-version field in plugin.yml.

ScriptCraft's Java code does not contain any deprecated calls so
it doesn't make much of a difference, but it still seems like
a good idea to declare the API we target.